LayoutTests/accessibility/aria-presentational-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<h3 role="presentation">
<a href="#">Link</a> and text
</h3>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria 'presentation' role works by successfully removing the element from the AX tree.");

    if (window.accessibilityController) {

          var body = document.getElementById("body");
          body.focus();

          var firstChild = accessibilityController.focusedElement.childAtIndex(0);
          shouldBe("firstChild.role", "'AXRole: AXLink'");
    }

    successfullyParsed = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
